+++ dev/src/write-batch.ts
@@ -29,7 +29,7 @@ import {FieldPath, validateFieldPath} from './path';
 import {validateDocumentReference} from './reference';
 import {Serializer, validateUserInput} from './serializer';
 import {Timestamp} from './timestamp';
-import {UpdateMap} from './types';
+import {FirestoreUnaryMethod, UpdateMap} from './types';
 import {
   getRetryCodes,
   isObject,
@@ -45,9 +45,8 @@ import {
   validateMinNumberOfArguments,
   validateOptional,
 } from './validate';
-
-import api = google.firestore.v1;
 import {GoogleError, Status} from 'google-gax';
+import api = google.firestore.v1;
 
 /**
  * A WriteResult wraps the write time set by the Firestore servers on sets(),
@@ -114,7 +113,7 @@ export class BatchWriteResult {
  * serializing the request.
  * @private
  */
-type PendingWriteOp = () => api.IWrite;
+export type PendingWriteOp = () => api.IWrite;
 
 /**
  * A Firestore WriteBatch that can be used to atomically commit multiple write
@@ -139,35 +138,19 @@ export class WriteBatch implements firestore.WriteBatch {
   private _committed = false;
 
   /**
-   * @hideconstructor
-   *
-   * @param firestore The Firestore Database client.
-   * @param retryBatch The WriteBatch that needs to be retried.
-   * @param indexesToRetry The indexes of the operations from the provided
-   * WriteBatch that need to be retried.
+   * The number of writes in this batch.
    */
-  constructor(
-    firestore: Firestore,
-    retryBatch: WriteBatch,
-    indexesToRetry: Set<number>
-  );
-  constructor(firestore: Firestore);
-  constructor(
-    firestore: Firestore,
-    retryBatch?: WriteBatch,
-    indexesToRetry?: Set<number>
-  ) {
+  get _opCount(): number {
+    return this._ops.length;
+  }
+
+  /**
+   * @hideconstructor
+   */
+  constructor(firestore: Firestore) {
     this._firestore = firestore;
     this._serializer = new Serializer(firestore);
     this._allowUndefined = !!firestore._settings.ignoreUndefinedProperties;
-
-    if (retryBatch) {
-      // Creates a new WriteBatch containing only the indexes from the provided
-      // indexes to retry.
-      for (const index of indexesToRetry!.values()) {
-        this._ops.push(retryBatch._ops[index]);
-      }
-    }
   }
 
   /**
@@ -494,9 +477,13 @@ export class WriteBatch implements firestore.WriteBatch {
         validateMaxNumberOfArguments('update', arguments, 3);
 
         const data = dataOrField as firestore.UpdateData;
-        Object.keys(data).forEach(key => {
-          validateFieldPath(key, key);
-          updateMap.set(FieldPath.fromArgument(key), data[key]);
+        Object.entries(data).forEach(([key, value]) => {
+          // Skip `undefined` values (can be hit if `ignoreUndefinedProperties`
+          // is set)
+          if (value !== undefined) {
+            validateFieldPath(key, key);
+            updateMap.set(FieldPath.fromArgument(key), value);
+          }
         });
 
         if (preconditionOrValues.length > 0) {
@@ -567,53 +554,23 @@ export class WriteBatch implements firestore.WriteBatch {
     // Capture the error stack to preserve stack tracing across async calls.
     const stack = Error().stack!;
 
-    return this._commit().catch(err => {
-      throw wrapError(err, stack);
-    });
-  }
+    // Commits should also be retried when they fail with status code ABORTED.
+    const retryCodes = [Status.ABORTED, ...getRetryCodes('commit')];
 
-  /**
-   * Commits all pending operations to the database and verifies all
-   * preconditions.
-   *
-   * The writes in the batch are not applied atomically and can be applied out
-   * of order.
-   *
-   * @private
-   */
-  async bulkCommit(): Promise<BatchWriteResult[]> {
-    this._committed = true;
-    const tag = requestTag();
-    await this._firestore.initializeIfNeeded(tag);
-
-    const database = this._firestore.formattedName;
-    const request: api.IBatchWriteRequest = {
-      database,
-      writes: this._ops.map(op => op.op()),
-    };
-
-    const retryCodes = getRetryCodes('batchWrite');
-
-    const response = await this._firestore.request<
-      api.IBatchWriteRequest,
-      api.BatchWriteResponse
-    >('batchWrite', request, tag, retryCodes);
-
-    return response.writeResults.map((result, i) => {
-      const status = response.status[i];
-      const error = new GoogleError(status.message || undefined);
-      error.code = status.code as Status;
-
-      // Since delete operations currently do not have write times, use a
-      // sentinel Timestamp value.
-      // TODO(b/158502664): Use actual delete timestamp.
-      const DELETE_TIMESTAMP_SENTINEL = Timestamp.fromMillis(0);
-      const updateTime =
-        error.code === Status.OK
-          ? Timestamp.fromProto(result.updateTime || DELETE_TIMESTAMP_SENTINEL)
-          : null;
-      return new BatchWriteResult(updateTime, error);
-    });
+    return this._commit<api.CommitRequest, api.CommitResponse>({retryCodes})
+      .then(response => {
+        return (response.writeResults || []).map(
+          writeResult =>
+            new WriteResult(
+              Timestamp.fromProto(
+                writeResult.updateTime || response.commitTime!
+              )
+            )
+        );
+      })
+      .catch(err => {
+        throw wrapError(err, stack);
+      });
   }
 
   /**
@@ -626,23 +583,25 @@ export class WriteBatch implements firestore.WriteBatch {
    * this request.
    * @returns  A Promise that resolves when this batch completes.
    */
-  async _commit(commitOptions?: {
+  async _commit<Req extends api.ICommitRequest, Resp>(commitOptions?: {
     transactionId?: Uint8Array;
     requestTag?: string;
-  }): Promise<WriteResult[]> {
+    retryCodes?: number[];
+    methodName?: FirestoreUnaryMethod;
+  }): Promise<Resp> {
     // Note: We don't call `verifyNotCommitted()` to allow for retries.
     this._committed = true;
 
-    const tag = (commitOptions && commitOptions.requestTag) || requestTag();
+    const tag = commitOptions?.requestTag ?? requestTag();
     await this._firestore.initializeIfNeeded(tag);
 
-    const database = this._firestore.formattedName;
-    const explicitTransaction = commitOptions && commitOptions.transactionId;
-
+    // Note that the request may not always be of type ICommitRequest. This is
+    // just here to ensure type safety.
     const request: api.ICommitRequest = {
-      database,
+      database: this._firestore.formattedName,
       writes: this._ops.map(op => op.op()),
     };
+
     if (commitOptions?.transactionId) {
       request.transaction = commitOptions.transactionId;
     }
@@ -654,26 +613,11 @@ export class WriteBatch implements firestore.WriteBatch {
       request.writes!.length
     );
 
-    let retryCodes: number[] | undefined;
-
-    if (explicitTransaction) {
-      request.transaction = explicitTransaction;
-    } else {
-      // Commits outside of transaction should also be retried when they fail
-      // with status code ABORTED.
-      retryCodes = [Status.ABORTED, ...getRetryCodes('commit')];
-    }
-
-    const response = await this._firestore.request<
-      api.ICommitRequest,
-      api.CommitResponse
-    >('commit', request, tag, retryCodes);
-
-    return (response.writeResults || []).map(
-      writeResult =>
-        new WriteResult(
-          Timestamp.fromProto(writeResult.updateTime || response.commitTime!)
-        )
+    return this._firestore.request<Req, Resp>(
+      commitOptions?.methodName || 'commit',
+      request as Req,
+      tag,
+      commitOptions?.retryCodes
     );
   }
 
@@ -959,16 +903,8 @@ function validateUpdateMap(
   if (!isPlainObject(obj)) {
     throw new Error(customObjectMessage(arg, obj));
   }
-
-  let isEmpty = true;
-  if (obj) {
-    for (const prop of Object.keys(obj)) {
-      isEmpty = false;
-      validateFieldValue(arg, obj[prop], allowUndefined, new FieldPath(prop));
-    }
-  }
-
-  if (isEmpty) {
+  if (Object.keys(obj).length === 0) {
     throw new Error('At least one field must be updated.');
   }
+  validateFieldValue(arg, obj, allowUndefined);
 }